Fix SCAN Limit wire-width mismatch with the fitz broker
KvScanQuery.Limit was declared ulong? and written on the wire with
WriteU64 (8 bytes), but the fitz broker's SCAN decoder
(read_optional_limit) only reads a u32 (4 bytes) -- matching
MAX_SCAN_ITEMS and every other broker-side limit. The extra 4 bytes
shifted every subsequent field (reverse, optional start_exclusive),
which the broker's decoder correctly flagged as trailing data:

  Cntryl.Fitz.KvException: SCAN failed: Invalid request: Trailing
  data in SCAN payload

This meant every KvDirectory<T,TKey>.QueryAsync/ScanAllAsync call
against a real broker failed -- point reads/writes were unaffected,
only Scan. Root-caused by rebuilding the fitz broker from its own
current main and reproducing the identical error locally, confirming
this is a genuine field-width bug rather than a stale broker image.

Fixes it by changing KvScanQuery.Limit to uint?/WriteU32 to match the
broker's u32, updating the three (ulong) call sites in KvDirectory
that constructed it, and the two tests that asserted on the old type.

+++ src/Fitz.Abstractions/Domains/Kv/KvTypes.cs
@@ -59,7 +59,7 @@ public sealed record KvScanResult(IReadOnlyList<KvPair> Pairs, bool HasMore);
 public sealed record KvScanQuery(
     ReadOnlyMemory<byte>? StartKey = null,
     ReadOnlyMemory<byte>? EndKey = null,
-    ulong? Limit = null,
+    uint? Limit = null,
     bool Reverse = false);
 
 /// <summary>A committed KV mutation notification.</summary>
